Changed Some Stuff
Bowtie, Quote of the Day, and Calendar changes. The Bowtie widget is
slightly messed up with Spotify (The progress bar and album art is
broken. Album art may be fixable but I need to look into it). It’s not
my fault. Spotify’s applescript API is slightly broken. They said they
are working on it and it should work as intended eventually.